+++ lib/react/auto-import.rb
@@ -6,13 +6,13 @@ if RUBY_ENGINE == 'opal'
   # to auto import native libraries and components using React::NativeLibrary
   class Object
     class << self
-      alias _reactrb_original_const_missing const_missing
-      alias _reactrb_original_method_missing method_missing
+      alias _react_original_const_missing const_missing
+      alias _react_original_method_missing method_missing
 
       def const_missing(const_name)
         # Opal uses const_missing to initially define things,
         # so we always call the original, and respond to the exception
-        _reactrb_original_const_missing(const_name)
+        _react_original_const_missing(const_name)
       rescue StandardError => e
         React::NativeLibrary.import_const_from_native(Object, const_name, true) || raise(e)
       end
@@ -20,7 +20,7 @@ if RUBY_ENGINE == 'opal'
       def method_missing(method, *args, &block)
         # ToDo: call import_const_from_natve only, if methods starts with capital letter
         component_class = React::NativeLibrary.import_const_from_native(self, method, false)
-        return _reactrb_original_method_missing(method, *args, &block) unless component_class
+        return _react_original_method_missing(method, *args, &block) unless component_class
         React::RenderingContext.render(component_class, *args, &block)
       end
     end