+++ lib/react/component/base.rb
@@ -4,7 +4,7 @@ module React
       def self.inherited(child)
         # note this is turned off during old style testing:  See the spec_helper
         unless child.to_s == "React::Component::HyperTestDummy"
-          React::Component.deprecation_warning child, "The class name React::Component::Base has been deprecated.  Use Hyperloop::Component instead."
+          React::Component.deprecation_warning child, "The class name React::Component::Base has been deprecated.  Use Hyperstack::Component instead."
         end
         child.include(ComponentNoNotice)
       end