+++ lib/react/component/class_methods.rb
@@ -7,7 +7,7 @@ module React
         React::Component.deprecation_warning(self, message)
       end
 
-      def reactrb_component?
+      def react_component?
         true
       end
 
@@ -42,10 +42,6 @@ module React
         )
       end
 
-      def validator
-        @validator ||= Validator.new(props_wrapper)
-      end
-
       def prop_types
         if self.validator
           {
@@ -58,49 +54,6 @@ module React
           }
         else
           {}
-        end
-      end
-
-      def default_props
-        validator.default_props
-      end
-
-      def params(&block)
-        validator.build(&block)
-      end
-
-      def props_wrapper
-        @props_wrapper ||= Class.new(PropsWrapper)
-      end
-
-      def param(*args)
-        if args[0].is_a? Hash
-          options = args[0]
-          name = options.first[0]
-          default = options.first[1]
-          options.delete(name)
-          options.merge!({default: default})
-        else
-          name = args[0]
-          options = args[1] || {}
-        end
-        if options[:default]
-          validator.optional(name, options)
-        else
-          validator.requires(name, options)
-        end
-      end
-
-      def collect_other_params_as(name)
-        validator.allow_undefined_props = true
-        validator_in_lexical_scope = validator
-        props_wrapper.define_method(name) do
-          @_all_others ||= validator_in_lexical_scope.undefined_props(props)
-        end
-
-        validator_in_lexial_scope = validator
-        props_wrapper.define_method(name) do
-          @_all_others ||= validator_in_lexial_scope.undefined_props(props)
         end
       end
 
@@ -173,7 +126,7 @@ module React
         end
       end
 
-      def to_n 
+      def to_n
         React::API.class_eval('@@component_classes')[self]
       end
     end