+++ CHANGELOG.md
@@ -23,11 +23,11 @@ Whitespace conventions:
 ### Added
 
 - `React::Server` is provided as a module wrapping the original `ReactDOMServer` API, require `react/server` to use it. (#186)
-- `React::Config` is introduced, `environment` is the only config option provided for now. See [#204](https://github.com/ruby-hyperloop/hyper-react/issues/204) for usage details.
+- `React::Config` is introduced, `environment` is the only config option provided for now. See [#204](https://github.com/ruby-hyperstack/hyper-react/issues/204) for usage details.
 
 ### Changed
 
-- State syntax is now consistent with Hyperloop::Store, old syntax is deprecated. (#209, #97)
+- State syntax is now consistent with Hyperstack::Store, old syntax is deprecated. (#209, #97)
 
 ### Deprecated